backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
ollama==0.1.6
pydantic==2.6.1
python-multipart==0.0.9
pypdf==4.0.1
python-dotenv==1.0.1
requests==2.31.0
firebase-admin==6.5.0
